Revert "For testing the PEXTRA DynVar MIP variables, these variables have been taken off from the ignored file. Note this needs to be done later if all is working in the pre-ignored file as well #143."
This reverts commit dc3fa81b0e2a73ba88ea6aae174c6aa9f45bc1c5.

We revert this to avoid later unintended conflicting merge, because at the same time the 01.00.29 identification will take place in the master. Instead we will upload a seperate copy of the
ignored file with the changes we now revert. This can be used by locally when testing to overwrite the versioned file. #143.